+++ ads-management/src/main/java/com/adsmanagement/districts/DistrictRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DistrictRepository extends JpaRepository<District, Integer> {
-    Page<District> findAllByCity_Id(Integer cityId, Pageable pageable);
+public interface DistrictRepository extends JpaRepository<District, Short> {
+    Page<District> findAllByCity_Id(Short cityId, Pageable pageable);
 
     Page<District> findAll(Pageable pageable);
